+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [v1.3.0] - 2022-07-08
 ### Fixed
 - Bug that pod metadata was not preserved when creating the pod ([#87](https://github.com/jthomperoo/custom-pod-autoscaler-operator/issues/87)).
 ### Changed
@@ -113,7 +115,8 @@ Autoscaler.
 - Allow creation/deletion of CPA.
 
 [Unreleased]:
-https://github.com/jthomperoo/custom-pod-autoscaler-operator/compare/v1.2.1...HEAD
+https://github.com/jthomperoo/custom-pod-autoscaler-operator/compare/v1.3.0...HEAD
+[v1.3.0]: https://github.com/jthomperoo/custom-pod-autoscaler-operator/compare/v1.2.1...v1.3.0
 [v1.2.1]: https://github.com/jthomperoo/custom-pod-autoscaler-operator/compare/v1.2.0...v1.2.1
 [v1.2.0]: https://github.com/jthomperoo/custom-pod-autoscaler-operator/compare/v1.1.1...v1.2.0
 [v1.1.1]: https://github.com/jthomperoo/custom-pod-autoscaler-operator/compare/v1.1.0...v1.1.1